frontend/requirements.txt
flask
torch
torchvision
pillow
werkzeug